Updating infix to use the helper constructs
