+++ include/trieste/trieste.h
@@ -29,7 +29,30 @@ namespace trieste
     return result;
   }
 
+
+  inline ProcessResult operator>>(Reader& reader, Rewriter&& rewriter)
+  {
+    ProcessResult result = reader.read();
+    if (result.ok)
+    {
+      return rewriter.rewrite(result.ast);
+    }
+
+    return result;
+  }
+
   inline ProcessResult operator>>(Reader& reader, Writer& writer)
+  {
+    ProcessResult result = reader.read();
+    if (result.ok)
+    {
+      return writer.write(result.ast);
+    }
+
+    return result;
+  }
+
+  inline ProcessResult operator>>(Reader& reader, Writer&& writer)
   {
     ProcessResult result = reader.read();
     if (result.ok)
@@ -51,7 +74,28 @@ namespace trieste
     return result;
   }
 
+  inline ProcessResult
+  operator>>(const ProcessResult& result, Rewriter&& rewriter)
+  {
+    if (result.ok)
+    {
+      return rewriter.rewrite(result.ast);
+    }
+
+    return result;
+  }
+
   inline ProcessResult operator>>(const ProcessResult& result, Writer& writer)
+  {
+    if (result.ok)
+    {
+      return writer.write(result.ast);
+    }
+
+    return result;
+  }
+
+  inline ProcessResult operator>>(const ProcessResult& result, Writer&& writer)
   {
     if (result.ok)
     {
@@ -66,7 +110,17 @@ namespace trieste
     return rewriter.rewrite(ast->clone());
   }
 
+  inline ProcessResult operator>>(const Node& ast, Rewriter&& rewriter)
+  {
+    return rewriter.rewrite(ast->clone());
+  }
+
   inline ProcessResult operator>>(const Node& ast, Writer& writer)
+  {
+    return writer.write(ast->clone());
+  }
+
+  inline ProcessResult operator>>(const Node& ast, Writer&& writer)
   {
     return writer.write(ast->clone());
   }
