+++ parsers/yaml/event_writer.cc
@@ -390,7 +390,7 @@ namespace trieste
 {
   namespace yaml
   {
-    Writer event_writer(const std::string& path, const std::string& newline)
+    Writer event_writer(const std::filesystem::path& path, const std::string& newline)
     {
       return Writer(
         "yaml_event",
