+++ samples/infix/CMakeLists.txt
@@ -1,14 +1,25 @@
-add_executable(infix
-  lang.cc
-  main.cc
+add_executable(infix_trieste
+  reader.cc
+  infix_trieste.cc
   parse.cc
   )
 
+target_link_libraries(infix_trieste
+  trieste::trieste
+)
+ 
+add_test(NAME infix COMMAND infix_trieste test -f)
+
+add_executable(infix
+  reader.cc
+  writers.cc
+  infix.cc
+  parse.cc
+)
+
 target_link_libraries(infix
   trieste::trieste
-  )
+)
 
-add_test(NAME infix COMMAND infix test -f)
-
-install(TARGETS infix RUNTIME DESTINATION infix)
+install(TARGETS infix infix_trieste RUNTIME DESTINATION infix)
 install(DIRECTORY examples DESTINATION infix)
