--- samples/infix/main.cc
@@ -1,6 +1,0 @@
-#include "lang.h"
-
-int main(int argc, char** argv)
-{
-  return infix::driver().run(argc, argv);
-}
